requirements.txt
kafka-python
textblob
streamlit
pymongo
